+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Order field mapping profile: Fix the values in the Payment status field (UIDATIMP-1386)
 * Order import: creating an order with Receipt not required causes an error (UIDATIMP-1389)
 * Data Import Action/Field mapping profiles in modal window are not in alpha order by default on Create screen (UIDATIMP-1390)
+* Field mapping profile Invoice: enabled prop set to false when the value exists for Accounting code field (UIDATIMP-1392)
 
 ## [6.0.0](https://github.com/folio-org/ui-data-import/tree/v6.0.0) (2023-02-24)